Move the attestation related code to different file.
