+++ stage0/src/lib.rs
@@ -38,12 +38,6 @@ use x86_64::{
 };
 use zerocopy::AsBytes;
 
-use alloc::{string::String, vec::Vec};
-use coset::{cbor::value::Value, cwt, iana, CborSerializable, CoseError};
-use hkdf::Hkdf;
-use p384::ecdsa::{signature::Signer, Signature, SigningKey, VerifyingKey};
-use rand_core::OsRng;
-
 mod acpi;
 mod acpi_tables;
 mod allocator;
@@ -59,14 +53,12 @@ mod pic;
 mod sev;
 mod smp;
 mod zero_page;
+mod dice_attestation;
 
 type Measurement = [u8; 32];
 
 // Reserve 128K for boot data structures that will outlive Stage 0.
 static BOOT_ALLOC: allocator::BumpAllocator<0x20000> = allocator::BumpAllocator::uninit();
-
-pub const ID_SALT: &str = "DICE_ID_SALT";
-pub const SIGNATURE_LENGTH: usize = 20;
 
 // Heap for short-term allocations. These allocations are not expected to outlive Stage 0.
 #[cfg_attr(not(test), global_allocator)]
@@ -125,104 +117,6 @@ pub unsafe fn jump_to_kernel(entry_point: VirtAddr, zero_page: usize) -> ! {
 pub fn sev_status() -> SevStatus {
     // Safety: we don't allow mutation and this is initialized in the bootstrap assembly.
     unsafe { SEV_STATUS }
-}
-
-/// Attestation related functions
-///
-/// Returns Signed Stage1 key and measurments
-struct Stage0Signer {
-    signing_key: SigningKey,
-}
-
-// Signer implementation.
-impl Stage0Signer {
-    fn sign(&self, data: &[u8]) -> Vec<u8> {
-        let signed_stage1_ca_verifying_key: Signature = self.signing_key.sign(data);
-        Vec::from(signed_stage1_ca_verifying_key.to_bytes().as_slice())
-    }
-}
-
-fn derive_id(ikm: &[u8], info: Option<&[u8]>) -> [u8; SIGNATURE_LENGTH] {
-    let hkdf = Hkdf::<Sha256>::new(Some(&(ID_SALT.as_bytes())), ikm);
-    let mut okm: [u8; SIGNATURE_LENGTH] = [0u8; SIGNATURE_LENGTH];
-    hkdf.expand(info.unwrap_or(&[]), &mut okm)
-        .expect("20 is a valid length for Sha256 to output");
-
-    okm
-}
-
-fn generate_ecdsa_keys(info: &str) -> (SigningKey, [u8; SIGNATURE_LENGTH]) {
-    let key = SigningKey::random(&mut OsRng);
-    let public_key = VerifyingKey::from(&key);
-    let key_id = derive_id(
-        public_key.to_encoded_point(false).as_bytes(),
-        Some(info.as_bytes()),
-    );
-    (key, key_id)
-}
-
-fn generate_stage1_ca_cwt(
-    stage0_eca_key: SigningKey,
-    stage0_cert_id_hex: String,
-) -> Result<Vec<u8>, CoseError> {
-    // Generate Stage 1 keys and Signer.
-    let info_str = "ID";
-    let stage1_ca_key = generate_ecdsa_keys(info_str);
-    let stage1_ca_verifying_key = VerifyingKey::from(&stage1_ca_key.0);
-    let stage0_signer = Stage0Signer {
-        signing_key: stage0_eca_key,
-    };
-
-    // Generate claims
-    let claims = cwt::ClaimsSetBuilder::new()
-        .issuer(stage0_cert_id_hex)
-        .subject(hex::encode(stage1_ca_key.1))
-        .cwt_id(Vec::from(stage1_ca_key.1))
-        // Add additional private-use claim.
-        .private_claim(
-            -4670552,
-            Value::Bytes(Vec::from(
-                stage1_ca_verifying_key.to_encoded_point(false).as_bytes(),
-            )),
-        )
-        .build();
-
-    let aad = b"";
-
-    // Build a `CoseSign1` object.
-    let protected = coset::HeaderBuilder::new()
-        .algorithm(iana::Algorithm::ES384)
-        .build();
-    let unprotected = coset::HeaderBuilder::new()
-        .key_id((*b"AsymmetricECDSA384").into())
-        .build();
-    let sign1 = coset::CoseSign1Builder::new()
-        .protected(protected)
-        .unprotected(unprotected)
-        .payload(claims.clone().to_vec()?)
-        .create_signature(aad, |data| stage0_signer.sign(data))
-        .build();
-    sign1.to_vec()
-}
-
-pub fn generate_stage1_attestation() {
-    // Generate Stage0 keys. This key will be used to sign Stage1
-    // measurement+config and the stage1_ca_key
-    let info_str = "ID";
-    let stage0_ca_key = generate_ecdsa_keys(info_str);
-    let stage0_ca_verifying_key = VerifyingKey::from(&stage0_ca_key.0);
-
-    // Call generate_attestation_report() to get 'stage0_ca_verifying_key' added to attestation report.
-    log::debug!(
-        "Stage0 Verification key: {}",
-        hex::encode(stage0_ca_verifying_key.to_encoded_point(false).as_bytes())
-    );
-
-    // Generate Stage1 CWT
-    let stage1_cwt = generate_stage1_ca_cwt(stage0_ca_key.0, hex::encode(stage0_ca_key.1));
-    if stage1_cwt.is_ok() {
-        // Call code that transmits the serialized cwt to Stage1
-    }
 }
 
 /// Entry point for the Rust code in the stage0 BIOS.
@@ -338,7 +232,7 @@ pub fn rust64_start(encrypted: u64) -> ! {
         })
         .unwrap_or_default();
 
-    generate_stage1_attestation();
+    dice_attestation::generate_stage1_attestation();
 
     let kernel_info = kernel::try_load_kernel_image(&mut fwcfg, zero_page.e820_table())
         .unwrap_or(kernel::KernelInfo::default());
